focus/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.9
# by the following command:
#
#    pip-compile requirements.in
#
amqp==5.1.1
    # via kombu
appnope==0.1.3
    # via ipython
asgiref==3.6.0
    # via django
asttokens==2.2.1
    # via stack-data
async-timeout==4.0.2
    # via redis
backcall==0.2.0
    # via ipython
billiard==3.6.4.0
    # via celery
brotlipy==0.7.0
    # via -r requirements.in
celery[redis]==5.2.7
    # via
    #   -r requirements.in
    #   celery-redbeat
celery-redbeat==2.0.0
    # via -r requirements.in
certifi==2023.5.7
    # via sentry-sdk
cffi==1.15.1
    # via brotlipy
click==8.1.3
    # via
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
click-didyoumean==0.3.0
    # via celery
click-plugins==1.1.1
    # via celery
click-repl==0.2.0
    # via celery
decorator==5.1.1
    # via ipython
defusedxml==0.7.1
    # via odfpy
diff-match-patch==20230430
    # via django-import-export
dj-database-url==2.0.0
    # via -r requirements.in
django==3.2.19
    # via
    #   -r requirements.in
    #   dj-database-url
    #   django-import-export
    #   django-js-reverse
    #   django-log-request-id
    #   django-model-utils
    #   djangorestframework
django-debreach==2.1.0
    # via -r requirements.in
django-import-export==3.2.0
    # via -r requirements.in
django-js-reverse==0.9.1
    # via -r requirements.in
django-log-request-id==2.1.0
    # via -r requirements.in
django-model-utils==4.3.1
    # via -r requirements.in
django-webpack-loader==1.8.1
    # via -r requirements.in
djangorestframework==3.14.0
    # via -r requirements.in
et-xmlfile==1.1.0
    # via openpyxl
executing==1.2.0
    # via stack-data
gunicorn==20.1.0
    # via -r requirements.in
ipython==8.13.2
    # via -r requirements.in
jedi==0.18.2
    # via ipython
kombu==5.2.4
    # via celery
markuppy==1.14
    # via tablib
matplotlib-inline==0.1.6
    # via ipython
odfpy==1.4.1
    # via tablib
openpyxl==3.1.2
    # via tablib
parso==0.8.3
    # via jedi
pexpect==4.8.0
    # via ipython
pickleshare==0.7.5
    # via ipython
prompt-toolkit==3.0.38
    # via
    #   click-repl
    #   ipython
psutil==5.9.5
    # via -r requirements.in
psycopg2==2.9.6
    # via -r requirements.in
ptyprocess==0.7.0
    # via pexpect
pure-eval==0.2.2
    # via stack-data
pycparser==2.21
    # via cffi
pygments==2.15.1
    # via ipython
python-dateutil==2.8.2
    # via celery-redbeat
python-decouple==3.8
    # via -r requirements.in
pytz==2023.3
    # via
    #   celery
    #   django
    #   djangorestframework
pyyaml==6.0
    # via tablib
redis==4.5.5
    # via
    #   celery
    #   celery-redbeat
sentry-sdk==1.22.2
    # via -r requirements.in
six==1.16.0
    # via
    #   asttokens
    #   click-repl
    #   python-dateutil
sqlparse==0.4.4
    # via django
stack-data==0.6.2
    # via ipython
tablib[html,ods,xls,xlsx,yaml]==3.4.0
    # via django-import-export
tenacity==8.2.2
    # via celery-redbeat
traitlets==5.9.0
    # via
    #   ipython
    #   matplotlib-inline
typing-extensions==4.5.0
    # via
    #   dj-database-url
    #   ipython
urllib3==1.26.15
    # via sentry-sdk
vine==5.0.0
    # via
    #   amqp
    #   celery
    #   kombu
wcwidth==0.2.6
    # via prompt-toolkit
whitenoise==6.4.0
    # via -r requirements.in
xlrd==2.0.1
    # via tablib
xlwt==1.3.0
    # via tablib
google-api-python-client=2.86.0

google-auth-oauthlib=1.0.0
# The following packages are considered to be unsafe in a requirements file:
# setuptools
